# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для управления пользователями и профилями",
        "description": "Пример API для создания, редактирования и удаления профилей пользователей.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/store": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Отправка уведомления",
                "description": "Отправка уведомления в реальном времени.",
                "operationId": "744e6295d60052ff813176857abf6c2c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "numChannel": {
                                        "type": "string",
                                        "example": "general"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Hello, this is a real-time notification!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Уведомление отправлено"
                    },
                    "400": {
                        "description": "Ошибка запроса"
                    }
                }
            }
        },
        "/sendMessage": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Отправка сообщения",
                "description": "Отправка сообщения.",
                "operationId": "121324c3329d87021f1b2af2f316e1f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Hello, world!"
                                    },
                                    "numChannel": {
                                        "type": "string",
                                        "example": "general"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Сообщение отправлено успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Message OK"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка отправки сообщения",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Message Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}